refactor: 서비스 인터페이스와 서비스 구현 분리 #10
